+++ src/netlist/netlist_writer/netlist_writer_manager.cpp
@@ -7,6 +7,7 @@
 
 #include <chrono>
 #include <fstream>
+#include <vector>
 
 namespace hal
 {
@@ -41,6 +42,15 @@ namespace hal
             ProgramOptions description;
             description.add("--write-hdl", "Write netlist to file.", {ProgramOptions::A_REQUIRED_PARAMETER});
             return description;
+        }
+
+        std::vector<std::string> registered_extensions()
+        {
+            std::vector<std::string> retval;
+            retval.reserve(m_extension_to_writer.size());
+            for (auto it : m_extension_to_writer)
+                retval.push_back(it.first);
+            return retval;
         }
 
         void register_writer(const std::string& name, const WriterFactory& writer_factory, const std::vector<std::string>& supported_file_extensions)